service1/requirements.txt
Flask==2.3.3
psycopg2-binary==2.9.7
requests==2.31.0
opentelemetry-distro==0.45b0
opentelemetry-instrumentation-flask==0.45b0
opentelemetry-instrumentation-psycopg2==0.45b0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-instrumentation==0.45b0